host/requirements.txt
websockets>=12.0
mss>=9.0.1
Pillow>=10.0
pyautogui>=0.9.54
python-dotenv>=1.0.0
numpy>=1.24.0
